complete base service

+++ posts-microservice/src/application/posts/use_cases/post/delete_post.py
@@ -1,6 +1,7 @@
 from src.domain import PostId
 from src.application.posts.uow import PostUoW
 from src.application.posts.dto import DeletePostDTO
+from src.application.posts.exceptions import UserIsNotPostCreator
 from src.application.common import (
     UseCase,
     UseCaseData
@@ -12,6 +13,7 @@ class DeletePostData(UseCaseData):
     Данные для удаления поста
     """
     post_id: str
+    creator_id: int
 
     class Config:
         frozen = True
@@ -29,6 +31,13 @@ class DeletePostUseCase(UseCase):
         self._uow = uow
 
     async def __call__(self, data: DeletePostData) -> DeletePostDTO:
+        post = await self._uow.post_repo.get_post_by_id(
+            post_id=PostId(value=data.post_id)
+        )
+
+        if post.creator_id.get_value != data.creator_id:
+            raise UserIsNotPostCreator()
+
         await self._uow.post_repo.delete_post(
             post_id=PostId(value=data.post_id)
         )
